pyproject.toml
[project]
name = "adk-sample-story-teller"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-adk>=1.26.0",
]
